Move lmnet/lmnet/environment.py to blueoil/environment.py
