+++ blueoil/cmd/export.py
@@ -21,7 +21,7 @@ import PIL
 import numpy as np
 import tensorflow as tf
 
-from lmnet import environment
+from blueoil import environment
 from blueoil.utils.image import load_image
 from blueoil.utils import config as config_util
 from blueoil.utils import executor
