+++ lmnet/tests/executor_tests/test_build_tfds.py
@@ -18,7 +18,7 @@ import pytest
 
 from blueoil.cmd.build_tfds import run
 from executor.train import run as train_run
-from lmnet import environment
+from blueoil import environment
 from blueoil.datasets.dataset_iterator import DatasetIterator
 from blueoil.datasets.tfds import TFDSClassification, TFDSObjectDetection
 from blueoil.utils import config as config_util
