+++ lmnet/tests/lmnet_tests/datasets_tests/test_open_images_v4.py
@@ -198,7 +198,7 @@ def test_custom_has_validation_open_images_v4_object_detection():
 
 
 if __name__ == '__main__':
-    from lmnet.environment import setup_test_environment
+    from blueoil.environment import setup_test_environment
     setup_test_environment()
 
     test_open_images_v4_classification()
