+++ lmnet/tests/lmnet_tests/networks_tests/classification_test/test_lmnet_quantize.py
@@ -18,7 +18,7 @@ import tensorflow as tf
 from easydict import EasyDict
 
 from executor.train import start_training
-from lmnet import environment
+from blueoil import environment
 from lmnet.common import Tasks
 from blueoil.networks.classification.lmnet_quantize import LmnetQuantize
 from blueoil.datasets.image_folder import ImageFolderBase
