+++ lmnet/tests/lmnet_tests/networks_tests/object_detection_tests/test_yolo_v2_quantize.py
@@ -18,7 +18,7 @@ import tensorflow as tf
 from easydict import EasyDict
 
 from executor.train import start_training
-from lmnet import environment
+from blueoil import environment
 from lmnet.common import Tasks
 from blueoil.networks.object_detection.yolo_v2_quantize import YoloV2Quantize
 from blueoil.datasets.pascalvoc_2007 import Pascalvoc2007
